backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "eventlet>=0.40.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "python-dotenv>=1.2.1",
]
